skip run-ghc.sh test if ghc is not available

+++ tests/run-ghc.sh
@@ -2,6 +2,11 @@
 
 set -ev
 
+if ! which ghc; then
+    echo there is no ghc
+    exit 137
+fi
+
 rm -rf $0.dir
 mkdir $0.dir
 cd $0.dir
